Extract serialization helpers to api/serialization
Move the six JSON-representation helpers out of the god-module
api/__init__.py into a focused api/serialization.py:

- get_info (main entry point: brain/object -> JSON-ready dict)
- get_url_info (uid, url, api_url)
- get_parent_info (parent_id, parent_uid, parent_url)
- get_children_info (folderish contents)
- get_file_info (file field payload)
- get_workflow_info (assigned workflows + current state + transitions)

All six names remain importable from senaite.jsonapi.api via explicit
re-exports at the bottom of __init__.py, so fieldmanagers.py (which
calls api.get_file_info and api.get_url_info) and any downstream users
keep working unchanged.

Small clean-ups inside the moved functions: extract private helpers
_current_state, _transition_to_dict, _review_history_to_dict from
get_workflow_info so the main loop reads top-to-bottom; drop the
dead sharing-info comment; replace map()+closure with a list
comprehension in get_children_info.

Drop now-unused imports from __init__.py: bika.lims.api.snapshot,
Products.ATContentTypes.utils.DT2dt, IFieldManager.
